Fix bug in displaying past and future slots
The publish_date on a slot is stored as a dateTime object, so comparing with Date.today would show slots marked out for 'today' as available.
This commit changes this, so that slots set for 'today' will not be shown to normal users and will show as 'past slots' to owners.